back/requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
torchvision
transformers
pillow
numpy
pandas
